+++ lib/internal/modules/cjs/helpers.js
@@ -112,6 +112,7 @@ function stripBOM(content) {
 const builtinLibs = [
   'assert',
   'async_hooks',
+  'async_local_storage',
   'buffer',
   'child_process',
   'cluster',
